+++ test/workspace/include/derived.h
@@ -67,8 +67,8 @@ public:
     int operator()();
     char &operator[](int);
 
-    void *operator new(long unsigned int);
-    void *operator new[](long unsigned int);
+    void *operator new(size_t);
+    void *operator new[](size_t);
     void operator delete(void *);
     void operator delete[](void *);